Add files via upload
Updated second data visualization. It has a bar graph that shows total counts and a tile grid of responses. Each tile represents one response given and the color of the tile represents the conceptual category that appears first in the response. The numbers represent how many conceptual categories were coded for that single response.